jjp-133: add username to users table, rename package model to entity
