+++ auth_app/src/main/java/com/blankfactor/auth/service/UserService.java
@@ -1,6 +1,6 @@
 package com.blankfactor.auth.service;
 
-import com.blankfactor.auth.model.User;
+import com.blankfactor.auth.entity.User;
 import com.blankfactor.auth.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
