+++ auth_app/src/main/resources/messages.properties
@@ -3,6 +3,11 @@ email.required=Email is required.
 email.invalidFormat=Please provide a valid email address.
 password.required=Password is required.
 password.length=Password must be between 8 and 25 characters long.
+confirmPassword.required=Confirm password is required.
+confirmPassword.length=Confirm password must be between 8 and 25 characters long.
+username.required=Username is required.
+username.length=Username must not exceed 25 characters.
+username.pattern=Please, provide a valid username.
 firstName.required=First name is required.
 firstName.minLength=First name must be at least 2 characters long.
 lastName.required=Last name is required.
